added a test route
