+++ routes/index.js
@@ -1,11 +1,10 @@
 const express = require('express');
 const router = express.Router();
-const { index } = require('./user-routes');
 
 // middleware
 router.use(express.json())
 
 // routes
-router.get('/user', index);
+router.use('/users', require('./user-routes'));
 
 module.exports = router;
